remove all disable eslint by refactoring the code again
